+++ server/src/Ogmios/Control/MonadWebSocket.hs
@@ -14,7 +14,6 @@ module Ogmios.Control.MonadWebSocket
     , SubProtocol
     , Headers
     , headers
-    , subProtocols
 
       -- * Constants
     , pingThreadDelay
@@ -44,17 +43,12 @@ class Monad m => MonadWebSocket (m :: Type -> Type) where
         :: Connection -> ByteString -> m ()
     acceptRequest
         :: PendingConnection
-        -> Maybe SubProtocol
         -> (Connection -> m a)
         -> m a
 
 headers :: PendingConnection -> Headers
 headers =
     WS.requestHeaders . WS.pendingRequest
-
-subProtocols :: PendingConnection -> [ByteString]
-subProtocols =
-    WS.getRequestSubprotocols . WS.pendingRequest
 
 instance MonadWebSocket IO where
     receive =
@@ -63,9 +57,8 @@ instance MonadWebSocket IO where
         WS.sendTextData
     close =
         WS.sendClose
-    acceptRequest pending sub action = do
-        let accept = WS.defaultAcceptRequest { WS.acceptSubprotocol = sub }
-        conn <- WS.acceptRequestWith pending accept
+    acceptRequest pending action = do
+        conn <- WS.acceptRequestWith pending WS.defaultAcceptRequest
         WS.withPingThread conn pingThreadDelay afterEachPing (action conn)
       where
         afterEachPing :: IO ()
@@ -78,9 +71,9 @@ instance MonadWebSocket m => MonadWebSocket (ReaderT env m) where
         lift . send conn
     close conn =
         lift . close conn
-    acceptRequest pending sub action = do
+    acceptRequest pending action = do
         env <- ask
-        lift $ acceptRequest pending sub (\conn -> runReaderT (action conn) env)
+        lift $ acceptRequest pending (\conn -> runReaderT (action conn) env)
 
 --
 -- Constants